# [TEST-1485] Optimized test processing