I'm not sure these are enough, but I think with the INDEX, optimized UPSERT, and several code performance tweaks, the updates should go quite a bit faster now. There is still a lot of time spent checking for blank lines repeatedly, which, as you start to check millions of lines of code start to become problematic. I'd like to try these changes out on larger production jobs for now, and see if this is good enough before creating more complicated data structures and routines.

I set up for profiling with about 3k tests changed (out of ~25k) from an existing DB with the new index added. The results:
## Before:
```
============================== 3158 passed, 2 skipped, 22738 deselected, 1 xfailed, 1 xpassed, 12 warnings in 2485.73 seconds ==============================
Wrote profile results to pytest2.lprof
     2546.31 real      2082.16 user       181.71 sys
     2546.31 real      2082.16 user       181.71 sys
(datarobot-6.1-dr) #[14:50:41] carson.gee@carson:~/code/dr1 $
python2 -m line_profiler pytest2.lprof
Timer unit: 1e-06 s

Total time: 130.034 s
File: /Users/carson.gee/code/pytest-testmon/testmon/process_code.py
Function: cover_subindented_multilines at line 116

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
   116                                           @profile
   117                                           def cover_subindented_multilines(lines, start, end, indent_threshold):
   118   2163291     754260.0      0.3      0.6      fingerprints = []
   119   2163291     753477.0      0.3      0.6      in_subindented_area = False
   120  20603101    7240552.0      0.4      5.6      while start < end - 1:
   121  18439810    5792808.0      0.3      4.5          start += 1
   122  18439810    6112798.0      0.3      4.7          line = lines[start]
   123  18439810   18114027.0      1.0     13.9          if blank_re.match(lines[start]):
   124   4851939    1434470.0      0.3      1.1              continue
   125
   126  13587871   79925018.0      5.9     61.5          curr_indent = get_indent_level(line)
   127  13587871    4566260.0      0.3      3.5          if curr_indent <= indent_threshold:
   128     12931       7845.0      0.6      0.0              fingerprints.append(line)
   129     12931       4936.0      0.4      0.0              in_subindented_area = True
   130                                                   else:
   131  13574940    4032435.0      0.3      3.1              if in_subindented_area:
   132      6794       4071.0      0.6      0.0                  fingerprints.append(GAP_MARKS[curr_indent - 1])
   133      6794       2284.0      0.3      0.0                  in_subindented_area = False
   134
   135   2163291     668357.0      0.3      0.5      if in_subindented_area:
   136      2656       1620.0      0.6      0.0          fingerprints.append(GAP_MARKS[0])
   137   2163291     618733.0      0.3      0.5      return fingerprints

Total time: 186.115 s
File: /Users/carson.gee/code/pytest-testmon/testmon/process_code.py
Function: gap_marks_until at line 140

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
   140                                           @profile
   141                                           def gap_marks_until(lines, start, end):
   142   2163291     770602.0      0.4      0.4      if start < len(lines):
   143   2163291   10154122.0      4.7      5.5          indent_threshold = get_indent_level(lines[start]) - 1
   144                                               else:
   145                                                   indent_threshold = 0
   146
   147   2163291  173688039.0     80.3     93.3      fingerprints = cover_subindented_multilines(lines, start, end, indent_threshold)
   148   2163291    1501789.0      0.7      0.8      return [GAP_MARKS[indent_threshold]] + fingerprints, end

Total time: 22.3926 s
File: /Users/carson.gee/code/pytest-testmon/testmon/process_code.py
Function: covered_unused_statement at line 151

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
   151                                           @profile
   152                                           def covered_unused_statement(start, end, coverage):
   153  20745109    5536379.0      0.3     24.7      while start <= end:
   154  18581818   11273559.0      0.6     50.3          if start in coverage:
   155    117700      37495.0      0.3      0.2              return True
   156  18464118    4964139.0      0.3     22.2          start += 1
   157   2163291     581024.0      0.3      2.6      return False

Total time: 406.174 s
File: /Users/carson.gee/code/pytest-testmon/testmon/process_code.py
Function: create_fingerprints at line 160

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
   160                                           @profile
   161                                           def create_fingerprints(afile, special_blocks, coverage):
   162    100750      41105.0      0.4      0.0      line_idx = 0
   163    100750      39336.0      0.4      0.0      result = []
   164  68941848   24251520.0      0.4      6.0      while line_idx < len(afile):
   165  68841098   21566937.0      0.3      5.3          line_idx += 1
   166  68841098   23601526.0      0.3      5.8          line = afile[line_idx - 1]
   167
   168  68841098   56371527.0      0.8     13.9          if blank_re.match(line):
   169  14409034    4158911.0      0.3      1.0              continue
   170
   171                                                   if (
   172  54432064   18976271.0      0.3      4.7              line_idx in special_blocks
   173   2464064    1612813.0      0.7      0.4              and line_idx not in coverage
   174   2280991     727028.0      0.3      0.2              and not covered_unused_statement(
   175   2280991   40477735.0     17.7     10.0                  line_idx + 1, special_blocks[line_idx], coverage
   176                                                       )
   177                                                   ):
   178   2163291     719024.0      0.3      0.2              fingerprints, line_idx = gap_marks_until(
   179   2163291  192904856.0     89.2     47.5                  afile, line_idx - 1, special_blocks[line_idx]
   180                                                       )
   181   2163291    1069991.0      0.5      0.3              result.extend(fingerprints)
   182                                                   else:
   183  52268773   19622526.0      0.4      4.8              result.append(line)
   184    100750      33039.0      0.3      0.0      return result

Total time: 584.329 s
File: /Users/carson.gee/code/pytest-testmon/testmon/testmon_core.py
Function: make_nodedata at line 317

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
   317                                               @profile
   318                                               def make_nodedata(self, measured_files, default=None):
   319      3162       4692.0      1.5      0.0          result = {}
   320    103912     122397.0      1.2      0.0          for filename, covered in measured_files.items():
   321    100750      41563.0      0.4      0.0              if default:
   322                                                           result[filename] = default
   323                                                       else:
   324    100750    2660907.0     26.4      0.5                  if os.path.exists(os.path.join(self.rootdir, filename)):
   325    100750     350712.0      3.5      0.1                      module = self.source_tree.get_file(filename)
   326    100750      40258.0      0.4      0.0                      result[filename] = encode_lines(
   327    100750      37713.0      0.4      0.0                          create_fingerprints(
   328    100750  581069146.0   5767.4     99.4                              module.lines, module.special_blocks, covered
   329                                                                   )
   330                                                               )
   331      3162       1164.0      0.4      0.0          return result

Total time: 23.4449 s
File: /Users/carson.gee/code/pytest-testmon/testmon/testmon_core.py
Function: write_node_data at line 344

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
   344                                               @profile
   345                                               def write_node_data(self, nodeid, nodedata, result={}, fake=False):
   346      3162       8646.0      2.7      0.0          with self.connection as con:
   347      3162      22574.0      7.1      0.1              failed = self.did_fail(result)
   348      3162      11748.0      3.7      0.1              cursor = con.cursor()
   349      3162       3067.0      1.0      0.0              cursor.execute(
   350                                                           """
   351                                                           INSERT OR REPLACE INTO node
   352                                                           (environment, name, result, failed)
   353                                                           VALUES (?, ?, ?, ?)
   354      3162       1408.0      0.4      0.0                  """,
   355      3162    4350012.0   1375.7     18.6                  (self.environment, nodeid, json.dumps(result), failed),
   356                                                       )
   357      3162      10573.0      3.3      0.0              node_id = cursor.lastrowid
   358
   359    103912    7082086.0     68.2     30.2              for filename in nodedata:
   360    100750      46782.0      0.5      0.2                  if fake:
   361                                                               mtime, checksum = None, None
   362                                                           else:
   363    100750     282347.0      2.8      1.2                      module = self.source_tree.get_file(filename)
   364    100750     125405.0      1.2      0.5                      mtime, checksum = module.mtime, module.checksum
   365
   366    100750    3469834.0     34.4     14.8                  fingerprint = checksums_to_blob(nodedata[filename])
   367    100750      61578.0      0.6      0.3                  con.execute(
   368                                                               """
   369                                                               INSERT OR IGNORE INTO fingerprint
   370                                                               (file_name, fingerprint, mtime, checksum)
   371                                                               VALUES (?, ?, ?, ?)
   372    100750      46189.0      0.5      0.2                      """,
   373    100750    3754811.0     37.3     16.0                      (filename, fingerprint, mtime, checksum),
   374                                                           )
   375
   376    100750      71814.0      0.7      0.3                  fingerprint_id, db_mtime, db_checksum = con.execute(
   377    100750      46337.0      0.5      0.2                      "SELECT id, mtime, checksum FROM fingerprint WHERE file_name = ? AND fingerprint=?",
   378    100750    2250426.0     22.3      9.6                      (filename, fingerprint,),
   379                                                           ).fetchone()
   380
   381                                                           if (
   382    100750     108864.0      1.1      0.5                      db_checksum != checksum or db_mtime != mtime
   383                                                           ):
   384                                                               self.update_mtimes([(mtime, checksum, fingerprint_id)])
   385
   386    100750      54556.0      0.5      0.2                  con.execute(
   387    100750      47265.0      0.5      0.2                      "INSERT INTO node_fingerprint VALUES (?, ?)",
   388    100750    1588620.0     15.8      6.8                      (node_id, fingerprint_id),
   389                                                           )

Total time: 0.181775 s
File: /Users/carson.gee/code/pytest-testmon/testmon/testmon_core.py
Function: sync_db_fs_nodes at line 391

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
   391                                               @profile
   392                                               def sync_db_fs_nodes(self, retain):
   393         1       4296.0   4296.0      2.4          collected = retain.union(self.stable_nodeids)
   394         1          4.0      4.0      0.0          with self.connection as con:
   395         1       4586.0   4586.0      2.5              add = collected - self.all_nodes
   396
   397         1          1.0      1.0      0.0              for nodeid in add:
   398                                                           if not is_python_file(home_file(nodeid)):
   399                                                               continue
   400                                                           self.write_node_data(
   401                                                               nodeid,
   402                                                               self.make_nodedata(
   403                                                                   {home_file(nodeid): None}, encode_lines(["0match"])
   404                                                               ),
   405                                                               fake=True,
   406                                                           )
   407
   408         1          2.0      2.0      0.0              con.executemany(
   409                                                           """
   410                                                           DELETE
   411                                                           FROM node
   412                                                           WHERE environment = ?
   413                                                             AND name = ?
   414         1          1.0      1.0      0.0                  """,
   415                                                           [
   416         1          0.0      0.0      0.0                      (self.environment, nodeid)
   417       374     172885.0    462.3     95.1                      for nodeid in self.all_nodes - collected
   418                                                           ],
   419                                                       )

Total time: 1.10422 s
File: /Users/carson.gee/code/pytest-testmon/testmon/testmon_core.py
Function: remove_unused_fingerprints at line 421

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
   421                                               @profile
   422                                               def remove_unused_fingerprints(self):
   423         1          2.0      2.0      0.0          with self.connection as con:
   424         1          0.0      0.0      0.0              con.execute(
   425                                                           """
   426                                                           DELETE FROM fingerprint
   427                                                           WHERE id NOT IN (
   428                                                               SELECT DISTINCT fingerprint_id FROM node_fingerprint
   429                                                           )
   430         1    1104218.0 1104218.0    100.0                  """
   431                                                       )
```

## After:
```
============================== 3158 passed, 2 skipped, 22738 deselected, 1 xfailed, 1 xpassed, 12 warnings in 2298.85 seconds ==============================
Wrote profile results to pytest2.lprof
     2365.08 real      1890.97 user       191.17 sys
     2365.08 real      1890.97 user       191.17 sys
(datarobot-6.1-dr) #[13:49:28] carson.gee@carson:~/code/dr1 $
python2 -m line_profiler pytest2.lprof
Timer unit: 1e-06 s

Total time: 64.8542 s
File: /Users/carson.gee/code/pytest-testmon/testmon/process_code.py
Function: cover_subindented_multilines at line 121

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
   121                                           @profile
   122                                           def cover_subindented_multilines(lines, start, end, indent_threshold):
   123   2163269     714223.0      0.3      1.1      fingerprints = []
   124   2163269     742738.0      0.3      1.1      in_subindented_area = False
   125  20598574    6871025.0      0.3     10.6      while start < end - 1:
   126  18435305    5829279.0      0.3      9.0          start += 1
   127  18435305    6104234.0      0.3      9.4          line = lines[start]
   128  18435305   17604355.0      1.0     27.1          if is_blank_line(lines[start]):
   129   4850789    1382724.0      0.3      2.1              continue
   130
   131  13584516   16068090.0      1.2     24.8          curr_indent = get_indent_level(line)
   132  13584516    4342665.0      0.3      6.7          if curr_indent <= indent_threshold:
   133     12921       6754.0      0.5      0.0              fingerprints.append(line)
   134     12921       4594.0      0.4      0.0              in_subindented_area = True
   135                                                   else:
   136  13571595    3897420.0      0.3      6.0              if in_subindented_area:
   137      6794       6006.0      0.9      0.0                  fingerprints.append(GAP_MARKS[curr_indent - 1])
   138      6794       2278.0      0.3      0.0                  in_subindented_area = False
   139
   140   2163269     665049.0      0.3      1.0      if in_subindented_area:
   141      2655       1859.0      0.7      0.0          fingerprints.append(GAP_MARKS[0])
   142   2163269     610940.0      0.3      0.9      return fingerprints

Total time: 115.558 s
File: /Users/carson.gee/code/pytest-testmon/testmon/process_code.py
Function: gap_marks_until at line 145

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
   145                                           @profile
   146                                           def gap_marks_until(lines, start, end):
   147   2163269     737569.0      0.3      0.6      if start < len(lines):
   148   2163269    2894451.0      1.3      2.5          indent_threshold = get_indent_level(lines[start]) - 1
   149                                               else:
   150                                                   indent_threshold = 0
   151
   152   2163269  110377009.0     51.0     95.5      fingerprints = cover_subindented_multilines(lines, start, end, indent_threshold)
   153   2163269    1549003.0      0.7      1.3      return [GAP_MARKS[indent_threshold]] + fingerprints, end

Total time: 7.12456 s
File: /Users/carson.gee/code/pytest-testmon/testmon/process_code.py
Function: covered_unused_statement at line 156

…